coords: ship cache/coords via git (float16, 44MB) — cluster cannot regenerate it
The cluster has no datasets/sky130hd/gates (2.8GB, the only table with cell
coordinates — every prior cache came from timing_paths/nets/pins), so
add_place_coords.py cannot run there, and rsync through the double jump host is
not workable. So this one cache ships in the repo.

Made it small enough to be reasonable: float32 -> float16 takes it 92MB -> 44MB.
Coords are normalized to [0,1] where float16 resolves ~5e-4 of the die, while
measured model geometry error is ~0.2 — quantization sits ~400x below the
signal. Verified end-to-end: targets load, stay in [0,1], VN boxes intact
(jpeg 349 VNs, box mean [0.391 0.373 0.588 0.621]).

.gitignore: cache/ -> cache/* plus !cache/coords/ (git cannot re-include a path
whose parent dir is excluded). Every other cache stays untracked — those ARE
regenerable in place from the dataset.

+++ scripts/add_place_coords.py
@@ -60,8 +60,13 @@ for di, d in enumerate(designs):
         W, H = max(x1 - x0, 1e-6), max(y1 - y0, 1e-6)
         xn = np.where(ok, (cx - x0) / W, 0.5)           # fill non-coord cells with die centre
         yn = np.where(ok, (cy - y0) / H, 0.5)
+        # float16, deliberately: this is the ONE cache the cluster cannot regenerate (the 2.8GB
+        # `gates` table is not there, and rsync through the double jump host is not workable), so
+        # it ships via git and has to be small. f32 -> 92MB, f16 -> 43MB. Coords are normalized
+        # to [0,1], where f16 resolves ~0.0005 of the die; model errors are ~0.2, so quantization
+        # sits 400x below the signal and costs nothing.
         np.savez_compressed(f"{OUT}/{fid}.npz",
-                            x=xn.astype(np.float32), y=yn.astype(np.float32),
+                            x=xn.astype(np.float16), y=yn.astype(np.float16),
                             mask=ok, die_w=np.float32(W), die_h=np.float32(H))
         tot_cells += ok.sum(); tot_missing += (~ok).sum(); tot_flows += 1
     print(f"[{di+1:2}/{len(designs)}] {d:14} {len(fids):4} flows", flush=True)
